fix: Inbound 코드 리뷰 피드백 반영 (검증 방식 @Valid + DTO로 통일, 소수점 유실 수정)

+++ src/main/java/com/app/maria/domain/inbound/dto/InboundDetailDTO.java
@@ -7,14 +7,13 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter @Setter @ToString @Builder
-
 public class InboundDetailDTO {
     private Long inboundDetailId;
     private Long inboundId;
     private Long foreignProductId;
     private String sourceBroker;
-    private Long qty;
-    private Long currentQty;
+    private BigDecimal qty;
+    private BigDecimal currentQty;
     private LocalDateTime recordedAt;
     private LocalDateTime purchaseDate;
     private BigDecimal purchasePrice;
